+++ climatoology/app/settings.py
@@ -28,6 +28,8 @@ class CABaseSettings(BaseSettings):
     postgres_password: str
     postgres_database: str
 
+    overwrite_assets: bool = False
+
     model_config = SettingsConfigDict(env_file='.env.base')
 
     @property